修正：液体机JEI显示错误 修改：方块机和液体机控制台输入修改为整数值 修改：版本号增加

+++ src/main/java/cn/sd/jrz/autoresource/blocks/BlockGeneratorBlock.java
@@ -112,7 +112,7 @@ public class BlockGeneratorBlock extends Block implements EntityBlock {
         if (generator == null) {
             return InteractionResult.FAIL;
         }
-        double block = generator.block / 1000D;
+        long block = generator.block / 1000;
         double output = generator.output / 1000D;
         double percent = (int) (generator.tickCount / 20.00D / generator.config.getSecond() * 10000D) / 100.00D;
         if (output < generator.config.getMax()) {
