+++ src/main/java/cn/sd/jrz/autoresource/items/LiquidGeneratorItem.java
@@ -29,7 +29,7 @@ public class LiquidGeneratorItem extends BlockItem {
     public void appendHoverText(@NotNull ItemStack stack, @Nullable Level worldIn, @NotNull List<Component> tooltip, @NotNull TooltipFlag flagIn) {
         super.appendHoverText(stack, worldIn, tooltip, flagIn);
         double output = config.getMin() / 1000D;
-        double liquid = 0;
+        long liquid = 0;
         long tickCount = 0;
         long second = config.getSecond();
         long step = config.getStep();
@@ -40,7 +40,7 @@ public class LiquidGeneratorItem extends BlockItem {
                     output = tag.getLong("output") / 1000D;
                 }
                 if (tag.contains("liquid", Tag.TAG_LONG)) {
-                    liquid = tag.getLong("liquid") / 1000D;
+                    liquid = tag.getLong("liquid") / 1000;
                 }
                 if (tag.contains("tickCount", Tag.TAG_LONG)) {
                     tickCount = tag.getLong("tickCount");
